refactor(dataTable): move uploadedAssetLog to UploadedAsset folder
